Minor Changes to Visit and Chart and GUI look
Visits are now inserted by date in medical record array list. chnaged
gui look in nurse and patient. added buttons for charts in patient gui.
Added input handler class which can be used to check inputs of doubles
to check for non numeric inputs. also checks if string is empty and
assigns values like 0 or "N/A" for convenience.